# .well-known/openapi.yaml
openapi: 3.0.3
info:
  title: LeeLens API
  description: |
    LeeLens 是一个基于 AI 的代码仓库智能解读平台，能够自动分析任意 GitHub 代码仓库，并生成结构化的项目文档。

    ## AI 友好设计

    本 API 针对以下 AI 工具进行了优化：
    - Claude Code (Anthropic)
    - GitHub Copilot
    - 其他支持 OpenAPI 的 AI 工具

    ## 核心功能

    - **仓库管理**：添加、删除、查看代码仓库
    - **任务管理**：执行、监控、重试分析任务
    - **文档管理**：查看、编辑、导出生成的文档
    - **API Key 管理**：配置多个 LLM API Key，支持自动切换
    - **数据同步**：与其他 LeeLens 实例同步数据
    - **用户需求管理**：提交和跟踪用户需求

    ## 响应格式

    所有 API 响应遵循统一格式：

    ### 成功响应
    ```json
    {
      "code": 0,
      "message": "success",
      "data": { ... }
    }
    ```

    ### 错误响应
    ```json
    {
      "error": "error message"
    }
    ```
  version: 1.0.0
  contact:
    name: LeeLens
    url: https://gitee.com/li-yuyanglee/eino-lee-lens
servers:
  - url: http://localhost:8080
    description: 开发环境
  - url: https://leelens.fly.dev
    description: 生产环境
...
paths:
...
  /api/api-keys:
    get:
      tags:
        - api-keys
      summary: 获取 API Key 列表
      description: 获取所有配置的 API Key（API Key 已脱敏）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/APIKey'
                  total:
                    type: integer
    post:
      tags:
        - api-keys
      summary: 创建 API Key
      description: 添加新的 LLM API Key 配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
  /api/api-keys/{id}:
    get:
      tags:
        - api-keys
      summary: 获取 API Key 详情
      description: 获取指定 API Key 的详细信息（API Key 已脱敏）
      parameters:
        - $ref: '#/components/parameters/APIKeyId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
    put:
      tags:
        - api-keys
      summary: 更新 API Key
      description: 更新 API Key 配置（不包括 API Key 本身）
      parameters:
        - $ref: '#/components/parameters/APIKeyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAPIKeyRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
    delete:
      tags:
        - api-keys
      summary: 删除 API Key
      description: 删除指定的 API Key 配置
      parameters:
        - $ref: '#/components/parameters/APIKeyId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: deleted successfully
  /api/api-keys/{id}/status:
    patch:
      tags:
        - api-keys
      summary: 更新 API Key 状态
      description: 启用或禁用 API Key
      parameters:
        - $ref: '#/components/parameters/APIKeyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [enabled, disabled]
                  description: 状态
              required:
                - status
      responses:
        '200':
          description: 状态已更新
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: status updated successfully
...
components:
  parameters:
...
    APIKeyId:
      name: id
      in: path
      required: true
      description: API Key ID
      schema:
        type: integer
  schemas:
...
    APIKey:
      type: object
      properties:
        id:
          type: integer
          description: API Key ID
        name:
          type: string
          description: API Key 名称
        provider:
          type: string
          description: 提供商
        base_url:
          type: string
          description: API 地址
        api_key:
          type: string
          description: API Key（脱敏）
        model:
          type: string
          description: 模型名称
        priority:
          type: integer
          description: 优先级
        status:
          type: string
          description: 状态
          enum: [enabled, disabled]
        request_count:
          type: integer
          description: 请求次数
        error_count:
          type: integer
          description: 错误次数
        created_at:
          type: string
          format: date-time
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          description: 更新时间
      required:
        - id
        - name
        - provider
        - base_url
        - model
    CreateAPIKeyRequest:
      type: object
      properties:
        name:
          type: string
          description: API Key 名称
        provider:
          type: string
          description: 提供商
        base_url:
          type: string
          description: API 地址
        api_key:
          type: string
          description: API Key
        model:
          type: string
          description: 模型名称
        priority:
          type: integer
          description: 优先级
      required:
        - name
        - provider
        - base_url
        - api_key
        - model
    UpdateAPIKeyRequest:
      type: object
      properties:
        name:
          type: string
          description: API Key 名称
        provider:
          type: string
          description: 提供商
        base_url:
          type: string
          description: API 地址
        model:
          type: string
          description: 模型名称
        priority:
          type: integer
          description: 优先级
...